ReviewMissionPlan is now based on MissionType instead of Class
Changes:
- ReviewMission task no longer uses Mission subclasses; scor ebased on
core Mission interface/abstract classes
- New interface SiteMission for Missions that use a site(s)
- Site based mission implement new interface
- ReviewMission uses ObectiveType -> MissionType(s) map to score
- Distance for review based on VehicleMission
- Removed GameMode from MissionManager as filtered is already handle in
UI client
- Mission in Console shows details of Mission Plan

#453

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/vehicle/Vehicle.java
@@ -895,7 +895,7 @@ public abstract class Vehicle extends Unit
 	 */
 	public double getRange(MissionType missionType) {
 		
-		int radius = (int)(getAssociatedSettlement().getMissionRadius(missionType));	
+		int radius = getAssociatedSettlement().getMissionRadius(missionType);	
 		double range = 0;
 		Mission mission = getMission();
